Move optimistic mutations logic into a separate file
